Alex

+++ src/main/java/com/citabot/interfaceService/IPacientePatologiaService.java
@@ -8,9 +8,9 @@ import java.util.Optional;
 public interface IPacientePatologiaService {
 
     public List<PacientePatologia> listar();
-    public List<PacientePatologia> listarByPacienteId(int id);
+    public List<PacientePatologia> listarByPacienteId(long id);
     public Optional<PacientePatologia> listarById(int id);
     public String delete(int id);
-    public PacientePatologia save(String tipo, int pacienteId, int enfermedadId);
+    public PacientePatologia save(String tipo, long pacienteId, int enfermedadId);
 
 }
